// src/app/utils/enums.ts
export enum InputTypes {
  INPUT = "input",
  SELECT = "select",
  CHECKBOX = "checkbox",
  DATE = "date",
  NUMBER = "number",
}

export enum InterceptorTypes {
  VALIDATION = "validation",
  CALCULATION = "calculation",
}

export enum EventTypes {
  ON_SUBMIT = "onSubmit",
  ON_CHANGE = "onChange",
  ON_KEY_DOWN = "onKeyDown",
  ON_LOAD = "onLoad",
}

export enum MessageTypes {
  ERROR = "error",
  INFO = "info",
}

export enum ComponentTypes {
  FORM = "form",
  INPUT_TABLE = "input-table",
}